requirements.txt
python-dotenv~=1.1.1
protobuf~=6.32.1
mcp~=1.14.1
pytest~=8.4.2
pytest-asyncio~=1.2.0
uvicorn~=0.36.0
fastapi~=0.116.2
starlette~=0.48.0
google-adk~=1.14.1
litellm~=1.77.1
pydantic~=2.11.9
a2a-sdk~=0.3.6
httpx~=0.28.1
sqlalchemy~= 2.0.43
pydantic[email]
elasticsearch~=8.17.0
psycopg2~=2.9.7
pymilvus~=2.5.6